+++ apps/backend/src/route-handlers/smart-request.tsx
@@ -1,11 +1,11 @@
 import "../polyfills";
 
-import { getUser } from "@/app/api/v1/users/crud";
-import { checkApiKeySet } from "@/lib/api-keys";
-import { getProject, listManagedProjectIds } from "@/lib/projects";
+import { getUser, getUserQuery } from "@/app/api/v1/users/crud";
+import { checkApiKeySet, checkApiKeySetQuery } from "@/lib/api-keys";
+import { getProjectQuery, listManagedProjectIds } from "@/lib/projects";
 import { decodeAccessToken } from "@/lib/tokens";
-import { traceSpan, withTraceSpan } from "@/utils/telemetry";
-import { Span } from "@opentelemetry/api";
+import { rawQueryAll } from "@/prisma-client";
+import { withTraceSpan } from "@/utils/telemetry";
 import { KnownErrors } from "@stackframe/stack-shared";
 import { ProjectsCrud } from "@stackframe/stack-shared/dist/interface/crud/projects";
 import { UsersCrud } from "@stackframe/stack-shared/dist/interface/crud/users";
@@ -13,7 +13,6 @@ import { StackAdaptSentinel, yupValidate } from "@stackframe/stack-shared/dist/s
 import { groupBy, typedIncludes } from "@stackframe/stack-shared/dist/utils/arrays";
 import { getNodeEnvironment } from "@stackframe/stack-shared/dist/utils/env";
 import { StackAssertionError, StatusError, throwErr } from "@stackframe/stack-shared/dist/utils/errors";
-import { ignoreUnhandledRejection } from "@stackframe/stack-shared/dist/utils/promises";
 import { deindent } from "@stackframe/stack-shared/dist/utils/strings";
 import { NextRequest } from "next/server";
 import * as yup from "yup";
@@ -139,7 +138,7 @@ export async function parseBody(req: NextRequest, bodyBuffer: ArrayBuffer): Prom
   }
 }
 
-const parseAuth = withTraceSpan('smart request parseAuth', async (req: NextRequest, span: Span): Promise<SmartRequestAuth | null> => {
+const parseAuth = withTraceSpan('smart request parseAuth', async (req: NextRequest): Promise<SmartRequestAuth | null> => {
   const projectId = req.headers.get("x-stack-project-id");
   let requestType = req.headers.get("x-stack-access-type");
   const publishableClientKey = req.headers.get("x-stack-publishable-client-key");
@@ -150,7 +149,7 @@ const parseAuth = withTraceSpan('smart request parseAuth', async (req: NextReque
   const refreshToken = req.headers.get("x-stack-refresh-token");
   const developmentKeyOverride = req.headers.get("x-stack-development-override-key");  // in development, the internal project's API key can optionally be used to access any project
 
-  const extractUserFromAccessToken = async (options: { token: string, projectId: string }) => {
+  const extractUserIdFromAccessToken = async (options: { token: string, projectId: string }) => {
     const result = await decodeAccessToken(options.token);
     if (result.status === "error") {
       throw result.error;
@@ -160,13 +159,7 @@ const parseAuth = withTraceSpan('smart request parseAuth', async (req: NextReque
       throw new KnownErrors.InvalidProjectForAccessToken();
     }
 
-    const user = await getUser({ projectId: options.projectId, userId: result.data.userId });
-    if (!user) {
-      // this is the case when access token is still valid, but the user is deleted from the database
-      throw new KnownErrors.AccessTokenExpired();
-    }
-
-    return user;
+    return result.data.userId;
   };
 
   const extractUserFromAdminAccessToken = async (options: { token: string, projectId: string }) => {
@@ -197,26 +190,18 @@ const parseAuth = withTraceSpan('smart request parseAuth', async (req: NextReque
     return user;
   };
 
-  // Do all the requests in parallel
 
-  const queryFuncs = {
-    project: () => projectId ? getProject(projectId) : Promise.resolve(null),
-    isClientKeyValid: () => projectId && publishableClientKey && requestType === "client" ? checkApiKeySet(projectId, { publishableClientKey }) : Promise.resolve(false),
-    isServerKeyValid: () => projectId && secretServerKey && requestType === "server" ? checkApiKeySet(projectId, { secretServerKey }) : Promise.resolve(false),
-    isAdminKeyValid: () => projectId && superSecretAdminKey && requestType === "admin" ? checkApiKeySet(projectId, { superSecretAdminKey }) : Promise.resolve(false),
-    user: () => projectId && accessToken ? extractUserFromAccessToken({ token: accessToken, projectId }) : Promise.resolve(null),
-    internalUser: () => projectId && adminAccessToken ? extractUserFromAdminAccessToken({ token: adminAccessToken, projectId }) : Promise.resolve(null),
-  } as const;
-  const results: [string, Promise<any>][] = [];
-  for (const [key, func] of Object.entries(queryFuncs)) {
-    results.push([
-      key,
-      ignoreUnhandledRejection(traceSpan(`auth query ${key}`, async () => {
-        return await func();
-      })),
-    ]);
-  }
-  const queries = Object.fromEntries(results) as { [K in keyof typeof queryFuncs]: ReturnType<typeof queryFuncs[K]> };
+  // Prisma does a query for every function call by default, even if we batch them with transactions
+  // Because smart route handlers are always called, we instead send over a single raw query that fetches all the
+  // data at the same time, saving us a lot of requests
+  const bundledQueries = {
+    user: projectId && accessToken ? getUserQuery(projectId, await extractUserIdFromAccessToken({ token: accessToken, projectId })) : undefined,
+    isClientKeyValid: projectId && publishableClientKey && requestType === "client" ? checkApiKeySetQuery(projectId, { publishableClientKey }) : undefined,
+    isServerKeyValid: projectId && secretServerKey && requestType === "server" ? checkApiKeySetQuery(projectId, { secretServerKey }) : undefined,
+    isAdminKeyValid: projectId && superSecretAdminKey && requestType === "admin" ? checkApiKeySetQuery(projectId, { superSecretAdminKey }) : undefined,
+    project: projectId ? getProjectQuery(projectId) : undefined,
+  };
+  const queriesResults = await rawQueryAll(bundledQueries);
 
   const eitherKeyOrToken = !!(publishableClientKey || secretServerKey || superSecretAdminKey || adminAccessToken);
 
@@ -234,30 +219,27 @@ const parseAuth = withTraceSpan('smart request parseAuth', async (req: NextReque
     const result = await checkApiKeySet("internal", { superSecretAdminKey: developmentKeyOverride });
     if (!result) throw new StatusError(401, "Invalid development key override");
   } else if (adminAccessToken) {
-    if (await queries.internalUser) {
-      if (!await queries.project) {
-        // this happens if the project is still in the user's managedProjectIds, but has since been deleted
-        throw new KnownErrors.InvalidProjectForAdminAccessToken();
-      }
-    } else {
-      // This case should be prevented by checks inside extractUserFromAdminAccessToken, if this happens, something is wrong
-      throw new StackAssertionError("adminAccessToken exists but no internal user was found");
+    // TODO put the assertion below into the bundled queries above (not so important because this path is quite rare)
+    await extractUserFromAdminAccessToken({ token: adminAccessToken, projectId });  // assert that the admin token is valid
+    if (!queriesResults.project) {
+      // this happens if the project is still in the user's managedProjectIds, but has since been deleted
+      throw new KnownErrors.InvalidProjectForAdminAccessToken();
     }
   } else {
     switch (requestType) {
       case "client": {
         if (!publishableClientKey) throw new KnownErrors.ClientAuthenticationRequired();
-        if (!await queries.isClientKeyValid) throw new KnownErrors.InvalidPublishableClientKey(projectId);
+        if (!queriesResults.isClientKeyValid) throw new KnownErrors.InvalidPublishableClientKey(projectId);
         break;
       }
       case "server": {
         if (!secretServerKey) throw new KnownErrors.ServerAuthenticationRequired();
-        if (!await queries.isServerKeyValid) throw new KnownErrors.InvalidSecretServerKey(projectId);
+        if (!queriesResults.isServerKeyValid) throw new KnownErrors.InvalidSecretServerKey(projectId);
         break;
       }
       case "admin": {
         if (!superSecretAdminKey) throw new KnownErrors.AdminAuthenticationRequired;
-        if (!await queries.isAdminKeyValid) throw new KnownErrors.InvalidSuperSecretAdminKey(projectId);
+        if (!queriesResults.isAdminKeyValid) throw new KnownErrors.InvalidSuperSecretAdminKey(projectId);
         break;
       }
       default: {
@@ -266,14 +248,14 @@ const parseAuth = withTraceSpan('smart request parseAuth', async (req: NextReque
     }
   }
 
-  const project = await queries.project;
+  const project = queriesResults.project;
   if (!project) {
     throw new StackAssertionError("Project not found; this should never happen because passing the checks until here should guarantee that the project exists and that access to it is granted", { projectId });
   }
 
   return {
     project,
-    user: await queries.user ?? undefined,
+    user: queriesResults.user ?? undefined,
     type: requestType,
   };
 });