+++ apps/dev-launchpad/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @stackframe/dev-launchpad
 
+## 2.7.4
+
+## 2.7.3
+
+### Patch Changes
+
+- Various changes
+
+## 2.7.2
+
 ## 2.7.1
 
 ## 2.7.0