+++ apps/e2e/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @stackframe/e2e-tests
 
+## 2.7.4
+
+### Patch Changes
+
+- @stackframe/stack-shared@2.7.4
+
+## 2.7.3
+
+### Patch Changes
+
+- Various changes
+- Updated dependencies
+  - @stackframe/stack-shared@2.7.3
+
+## 2.7.2
+
+### Patch Changes
+
+- Various changes
+- Updated dependencies
+  - @stackframe/stack-shared@2.7.2
+
 ## 2.7.1
 
 ### Patch Changes