+++ packages/stack-emails/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @stackframe/stack-emails
 
+## 2.7.4
+
+### Patch Changes
+
+- @stackframe/stack-shared@2.7.4
+- @stackframe/stack-ui@2.7.4
+
+## 2.7.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.7.3
+  - @stackframe/stack-ui@2.7.3
+
+## 2.7.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.7.2
+  - @stackframe/stack-ui@2.7.2
+
 ## 2.7.1
 
 ### Patch Changes