+++ packages/stack-shared/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @stackframe/stack-shared
 
+## 2.7.4
+
+### Patch Changes
+
+- @stackframe/stack-sc@2.7.4
+
+## 2.7.3
+
+### Patch Changes
+
+- Various changes
+  - @stackframe/stack-sc@2.7.3
+
+## 2.7.2
+
+### Patch Changes
+
+- Various changes
+  - @stackframe/stack-sc@2.7.2
+
 ## 2.7.1
 
 ### Patch Changes