+++ packages/stack/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @stackframe/stack
 
+## 2.7.4
+
+### Patch Changes
+
+- Various changes
+  - @stackframe/stack-sc@2.7.4
+  - @stackframe/stack-shared@2.7.4
+  - @stackframe/stack-ui@2.7.4
+
+## 2.7.3
+
+### Patch Changes
+
+- Various changes
+- Updated dependencies
+  - @stackframe/stack-shared@2.7.3
+  - @stackframe/stack-ui@2.7.3
+  - @stackframe/stack-sc@2.7.3
+
+## 2.7.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.7.2
+  - @stackframe/stack-ui@2.7.2
+  - @stackframe/stack-sc@2.7.2
+
 ## 2.7.1
 
 ### Patch Changes